+++ src/extensions/mvs/tree/mvs/mvs-builder.ts
@@ -1,14 +1,14 @@
 /**
- * Copyright (c) 2023 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2023-2024 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Adam Midlik <midlik@gmail.com>
  */
 
 import { deepClone, pickObjectKeys } from '../../../../mol-util/object';
 import { MVSData } from '../../mvs-data';
-import { ParamsOfKind, SubTreeOfKind } from '../generic/tree-schema';
+import { AdditionalProperties, ParamsOfKind } from '../generic/tree-schema';
 import { MVSDefaults } from './mvs-defaults';
-import { MVSKind, MVSNode, MVSTree, MVSTreeSchema } from './mvs-tree';
+import { MVSKind, MVSNode, MVSSubtree, MVSTree, MVSTreeSchema } from './mvs-tree';
 
 
 /** Create a new MolViewSpec builder containing only a root node. Example of MVS builder usage:
@@ -30,7 +30,7 @@ export function createMVSBuilder() {
 class _Base<TKind extends MVSKind> {
     protected constructor(
         protected readonly _root: Root,
-        protected readonly _node: SubTreeOfKind<MVSTree, TKind>,
+        protected readonly _node: MVSSubtree<TKind>,
     ) { }
     /** Create a new node, append as child to current _node, and return the new node */
     protected addChild<TChildKind extends MVSKind>(kind: TChildKind, params: ParamsOfKind<MVSTree, TChildKind>) {
@@ -38,10 +38,27 @@ class _Base<TKind extends MVSKind> {
         const node = {
             kind,
             params: pickObjectKeys(params, allowedParamNames) as unknown,
-        } as SubTreeOfKind<MVSTree, TChildKind>;
+        } as MVSSubtree<TChildKind>;
         this._node.children ??= [];
         this._node.children.push(node);
         return node;
+    }
+    /** Adds provided key-value pairs as additional properties to this node. Use value `null` to remove a property. */
+    additionalProperties(props: AdditionalProperties) {
+        this._node.additional_properties ??= {};
+        for (const key in props) {
+            const value = props[key];
+            if (value === undefined) {
+                // Do nothing, this should be equivalent to `key` not being there
+            } else if (value === null) {
+                // Remove property
+                delete this._node.additional_properties[key];
+            } else {
+                // Add property
+                this._node.additional_properties[key] = value;
+            }
+        }
+        return this;
     }
 }
 
@@ -49,7 +66,7 @@ class _Base<TKind extends MVSKind> {
 /** MVS builder pointing to the 'root' node */
 export class Root extends _Base<'root'> {
     constructor() {
-        const node: MVSNode<'root'> = { kind: 'root' };
+        const node: MVSNode<'root'> = { kind: 'root', params: {}, additional_properties: undefined };
         super(undefined as any, node);
         (this._root as Root) = this;
     }
@@ -236,6 +253,7 @@ export function builderDemo() {
     const struct = builder.download({ url: 'https://www.ebi.ac.uk/pdbe/entry-files/download/1og2_updated.cif' }).parse({ format: 'mmcif' }).modelStructure();
     struct.component().representation().color({ color: 'white' });
     struct.component({ selector: 'ligand' }).representation({ type: 'ball_and_stick' })
+        .additionalProperties({ repr_quality: 'high' })
         .color({ color: '#555555' })
         .color({ selector: { type_symbol: 'N' }, color: '#3050F8' })
         .color({ selector: { type_symbol: 'O' }, color: '#FF0D0D' })