+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Note that since we don't clearly distinguish between a public and private interf
 
 ## [Unreleased]
 
+- MolViewSpec: Support for transparency and additional properties
+
 ## [v4.7.1] - 2024-09-30
 
 - Improve `resolutionMode` (#1279)